Decoupled Debugger and Lib, and Debugger no longer includes src of lib
