+++ package.json
@@ -16,7 +16,7 @@
   "module": "dist/wasmboy.esm.js",
   "browser": "dist/wasmboy.umd.js",
   "scripts": {
-    "start": "npm run core:build && npx concurrently --kill-others --names \"DEBUGGER,CORE,LIB\" -c \"bgBlue.bold,bgMagenta.bold,bgGreen.bold\" \"npm run debugger:watch\" \"npm run core:watch\" \"npm run lib:watch\"",
+    "start": "npm run core:build && npm run lib:build && npx concurrently --kill-others --names \"DEBUGGER,CORE,LIB\" -c \"bgBlue.bold,bgMagenta.bold,bgGreen.bold\" \"npm run debugger:watch\" \"npm run core:watch\" \"npm run lib:watch\"",
     "dev": "npm run start",
     "watch": "npm run start",
     "build": "npm run core:build && npm run lib:build",
@@ -25,9 +25,9 @@
     "test:accuracy": "npm run build && npx mocha test/accuracy/accuracy-test.js --exit",
     "test:perf": "npm run test:performance",
     "test:performance": "npm run build && npx mocha test/performance/performance-test.js --exit",
-    "debugger:watch": "preact watch",
+    "debugger:watch": "preact watch --src debugger",
     "debugger:serve": "npm run debugger:build && preact serve",
-    "debugger:build": "preact build --no-prerender",
+    "debugger:build": "preact build --src debugger --no-prerender",
     "debugger:deploy": "npm run debugger:build && echo 'wasmboy.app' > build/CNAME && npx gh-pages -d build",
     "core:watch": "npx watch \"npm run core:build\" core",
     "core:build": "npx asc core/index.ts -b dist/core/index.untouched.wasm -t dist/core/index.untouched.wast -O3 --validate --sourceMap assets/index.untouched.wasm.map --memoryBase 0 && cp dist/core/index.untouched.wasm.map assets/index.untouched.wasm.map && echo \"Built Wasm!\"",
